# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gonemaster API
  version: 0.1.0
  description: |
    REST API for running Gonemaster engine tests with queueing, streaming, and cancellation.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development (API base)
...
paths:
...
  /tags:
    get:
      tags: [Tags]
      summary: List tags
      operationId: listTags
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Tag array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Tags]
      summary: Create a tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          description: Tag already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /analysis/cohorts/{id}:
    get:
      tags: [Analysis]
      summary: Get an analysis cohort
      operationId: getAnalysisCohort
      parameters:
        - $ref: '#/components/parameters/IDPath'
      responses:
        '200':
          description: Cohort
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisCohort'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Analysis]
      summary: Patch an analysis cohort
      operationId: patchAnalysisCohort
      parameters:
        - $ref: '#/components/parameters/IDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisCohortPatchRequest'
      responses:
        '200':
          description: Updated cohort
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisCohort'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Analysis]
      summary: Delete an analysis cohort
      operationId: deleteAnalysisCohort
      parameters:
        - $ref: '#/components/parameters/IDPath'
      responses:
        '204':
          description: Cohort deleted
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Admin token. Required on /api/v1 endpoints when the server is started
        with admin tokens configured (token mode); ignored in open mode. Clients
        send "Authorization: Bearer <token>"; the admin UI uses a session cookie
        set via POST /session.
  parameters:
    IDPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Tag:
      type: object
      required: [name, created_at, domain_count]
      properties:
        name:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
        domain_count:
          type: integer
        default_profile_id:
          type: integer
          format: int64
    TagCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
...
    AnalysisCohort:
      type: object
      required: [id, source_type, source_tag, label, analysis_enabled, public_enabled, is_default, sort_order, materialization_status, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        source_type:
          type: string
          enum: [tag]
        source_tag:
          type: string
        label:
          type: string
        description:
          type: string
        analysis_enabled:
          type: boolean
        public_enabled:
          type: boolean
        is_default:
          type: boolean
        sort_order:
          type: integer
        materialization_status:
          type: string
          enum: [pending, ready, failed]
        materialization_done:
          type: integer
        materialization_total:
          type: integer
        last_materialized_at:
          type: string
          format: date-time
        last_materialization_error:
          type: string
        default_snapshot_policy:
          type: string
          enum: [auto_latest, pinned]
        default_snapshot_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AnalysisCohortPatchRequest:
      type: object
      properties:
        label:
          type: string
        description:
          type: string
        analysis_enabled:
          type: boolean
        public_enabled:
          type: boolean
        is_default:
          type: boolean
        sort_order:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'